moved all featureSegments classes to statistics module and enable feature for internal build
