+++ app/build.gradle
@@ -346,7 +346,6 @@ dependencies {
     implementation "androidx.work:work-multiprocess:_"
     androidTestImplementation AndroidX.work.testing
     implementation AndroidX.work.rxJava2
-    testImplementation AndroidX.work.testing
 
     // Dagger
     kapt Google.dagger.compiler
