+++ app/src/main/java/com/duckduckgo/app/pixels/AppPixelName.kt
@@ -247,6 +247,4 @@ enum class AppPixelName(override val pixelName: String) : Pixel.PixelName {
     REMOTE_MESSAGE_SECONDARY_ACTION_CLICKED("m_remote_message_secondary_action_clicked"),
 
     CREATE_BLOOM_FILTER_ERROR("m_create_bloom_filter_error"),
-
-    DAILY_USER_EVENT_SEGMENT("m_daily_user_event_segment"),
 }
